Restrict Item deletion to authorized users

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -54,10 +54,17 @@ const Mutations = {
       `{
       id
       title
+      user{id}
     }`
     );
     // 2. Check if they own that item, or have permissions
-    // TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some((permission) =>
+      ["ADMIN", "ITEMDELETE"].includes(permission)
+    );
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permission to do that!");
+    }
     // 3. Delete item
     return ctx.db.mutation.deleteItem({ where }, info);
   },
